requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
pandas==2.2.1
python-dotenv==1.0.1
streamlit==1.33.0
plotly==5.22.0
